+++ frontend/.next/server/webpack-runtime.js
@@ -125,7 +125,7 @@
 /******/ 	
 /******/ 	/* webpack/runtime/getFullHash */
 /******/ 	(() => {
-/******/ 		__webpack_require__.h = () => ("df993c3b25061812")
+/******/ 		__webpack_require__.h = () => ("f800c3bb2b59004b")
 /******/ 	})();
 /******/ 	
 /******/ 	/* webpack/runtime/hasOwnProperty shorthand */