+++ frontend/.next/static/chunks/webpack.js
@@ -192,7 +192,7 @@
 /******/ 	
 /******/ 	/* webpack/runtime/getFullHash */
 /******/ 	!function() {
-/******/ 		__webpack_require__.h = function() { return "c2282d5fb333e84a"; }
+/******/ 		__webpack_require__.h = function() { return "61c830f96a6752c2"; }
 /******/ 	}();
 /******/ 	
 /******/ 	/* webpack/runtime/global */